requirements.txt
Flask
cx_Oracle
pandas
scikit-learn
numpy
python-dotenv
